Rename `_helpers/rename` to `case-style`
