+++ src/test/case-style.spec.ts
@@ -1,5 +1,5 @@
 import { assert } from "chai";
-import { CaseStyle, detectCaseStyle, join, rename, split } from "../../lib/_helpers/rename";
+import { CaseStyle, detectCaseStyle, join, rename, split } from "../lib/case-style";
 
 describe("rename", function () {
   describe("detectCaseStyle", function () {
