+++ src/test/circular-types/fs-tree/fs-node-type.ts
@@ -1,4 +1,4 @@
-import { CaseStyle } from "../../../lib/_helpers/rename";
+import { CaseStyle } from "../../../lib/case-style";
 import { SimpleEnumType } from "../../../lib/types/simple-enum";
 
 export enum FsNodeType {
