+++ src/test/types/try-union.spec.ts
@@ -1,5 +1,5 @@
 import { assert } from "chai";
-import { CaseStyle } from "../../lib/_helpers/rename";
+import { CaseStyle } from "../../lib/case-style";
 import { DocumentType } from "../../lib/types/document";
 import { IntegerType } from "../../lib/types/integer";
 import { TryUnionType } from "../../lib/types/try-union";
